Refactor ContractData lookup
